# Conflicts:
#	src/main/java/pl/edu/agh/cinema/CinemaApplication.java
#	src/main/java/pl/edu/agh/cinema/model/movie/Movie.java
#	src/main/java/pl/edu/agh/cinema/model/room/Room.java
#	src/main/java/pl/edu/agh/cinema/ui/MainController.java
#	src/main/java/pl/edu/agh/cinema/ui/userManager/editUserDialog/UserDialogFieldController.java
#	src/main/resources/fxml/userManager/editMovieDialog/userDialogFields.fxml

+++ src/main/java/pl/edu/agh/cinema/model/room/Room.java
@@ -15,7 +15,7 @@ public class Room {
     @Id
     @GeneratedValue
     @Getter
-    private Long id;
+    private int id;
 
     @Getter
     private String roomName;
@@ -50,12 +50,8 @@ public class Room {
 
     public Room(String roomName, int seatsNumber) {
         this();
+
         this.roomName=roomName;
         this.seatsNumber=seatsNumber;
     }
-
-//    @Override
-//    public String toString() {
-//        return roomName;
-//    }
 }
